Changed Double to BigDecimal to correctly support currency values.

+++ app/src/main/java/no/mhl/clarence/data/model/CurrencyValue.kt
@@ -1,7 +1,9 @@
 package no.mhl.clarence.data.model
 
+import java.math.BigDecimal
+
 
 data class CurrencyValue(
     val name: String,
-    val value: Double
+    val value: BigDecimal
 )
