+++ app/src/main/java/no/mhl/clarence/data/model/Latest.kt
@@ -1,7 +1,9 @@
 package no.mhl.clarence.data.model
 
+import java.math.BigDecimal
+
 data class Latest(
-    val rates: Map<String, Double>,
+    val rates: Map<String, BigDecimal>,
     val base: String,
     val date: String
 )
@@ -10,6 +12,6 @@ data class Latest(
 fun mapToRate(latest: Latest): Rate =
     Rate(0, latest.base, latest.date, mapToCurrencyValues(latest.rates))
 
-private fun mapToCurrencyValues(map: Map<String, Double>): List<CurrencyValue> =
+private fun mapToCurrencyValues(map: Map<String, BigDecimal>): List<CurrencyValue> =
     map.map { CurrencyValue(it.key, it.value) }
 // endregion
